backend/requirements.txt
flask
flask-cors
sqlalchemy
pymysql
python-dotenv
cryptography
PyJWT